requirements.txt
tensorflow
django
djangorestframework
numpy
pandas
scikit-learn
datasets
